rename QNode -> BaseQNode
