+++ pennylane/qnode_new/cv.py
@@ -24,7 +24,7 @@ import numpy as np
 import pennylane as qml
 from pennylane.operation import ObservableReturnTypes
 from pennylane.qnode_new.jacobian import JacobianQNode
-from pennylane.qnode_new.qnode import QuantumFunctionError
+from pennylane.qnode_new.base import QuantumFunctionError
 
 
 class CVQNode(JacobianQNode):
