+++ tests/qnode_new/test_interfaces_autograd.py
@@ -21,7 +21,7 @@ import pytest
 import numpy as np
 
 import pennylane as qml
-from pennylane.qnode_new.qnode import QuantumFunctionError
+from pennylane.qnode_new.base import QuantumFunctionError
 from pennylane.qnode_new.qubit import QubitQNode
 from pennylane.qnode_new.cv import CVQNode
