+++ tests/qnode_new/test_qnode_cv.py
@@ -20,7 +20,7 @@ import numpy as np
 import pennylane as qml
 from pennylane._device import Device
 from pennylane.operation import CVObservable
-from pennylane.qnode_new.qnode import QuantumFunctionError
+from pennylane.qnode_new.base import QuantumFunctionError
 from pennylane.qnode_new.cv import CVQNode
