+++ tests/qnode_new/test_qnode_metric_tensor.py
@@ -20,7 +20,7 @@ from scipy.linalg import block_diag
 
 import pennylane as qml
 from pennylane.qnode_new.qubit import QubitQNode
-from pennylane.qnode_new.qnode import QuantumFunctionError
+from pennylane.qnode_new.base import QuantumFunctionError
 from pennylane.plugins.default_qubit import Y, Z
